chore(scap): move scap_test.h to engine/test_input
